pyproject.toml
[tool.poetry]
name = "localsend"
version = "0.1.0"
description = ""
authors = ["RuslanUC <dev_ruslan_uc@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
muffin = "^0.101.0"
uvicorn = "^0.30.3"
httpx = "^0.27.0"
pyopenssl = "^24.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

